pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gdynia-thermal-db"
version = "0.1.0"
description = "Research-grade geospatial data engineering for the Gdynia thermal map project"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["geospatial", "thermal", "gdynia", "geopandas", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "pandas>=2.0",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyproj>=3.6",
    "requests>=2.31",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "mercantile>=1.2",
    "rioxarray>=0.15",
    "xarray>=2024.1",
    "rasterio>=1.3",
    "pyarrow>=14.0",
    "tqdm>=4.66",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "responses>=0.25",
    "pre-commit>=3.7",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.29",
    "notebook>=7.0",
]

[project.scripts]
gdynia-thermal = "gdynia_thermal_db.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
